+++ src/utils/audioController.js
@@ -4,6 +4,7 @@ import { __reciter, __translationReciter, __playbackSpeed, __audioSettings, __au
 import { staticEndpoint, wordsAudioURL } from '$data/websiteSettings';
 import { selectableReciters, selectableTranslationReciters, selectablePlaybackSpeeds, selectableAudioDelays } from '$data/options';
 import { fetchAndCacheJson } from '$utils/fetchData';
+import { checkOnlineAndAlert } from '$utils/offlineModeHandler';
 
 // Getting the audio element
 let audio = document.querySelector('#player');
@@ -106,7 +107,9 @@ export async function playVerseAudio(props) {
 }
 
 // Function to play word audio
-export function playWordAudio(props) {
+export async function playWordAudio(props) {
+	if (!(await checkOnlineAndAlert())) return;
+
 	resetAudioSettings();
 
 	const audioSettings = get(__audioSettings);
@@ -249,7 +252,7 @@ async function wordHighlighter() {
 			lastPlayedKey = audioSettings.playingWordKey;
 		}
 	} catch (error) {
-		console.warn('wordHighlighter error:', error);
+		console.warn(error);
 	}
 }
 
@@ -405,9 +408,6 @@ export function prepareVersesToPlay(key) {
 			// Set verses to play from the startVerse to the end of the chapter
 			setVersesToPlay({ location: 'verseOptionsOrModal', chapter, startVerse, endVerse: endVerse });
 			break;
-		default:
-			// Handle invalid audioRange values
-			console.error('Invalid audioRange:', audioRange);
 	}
 }
 
@@ -427,6 +427,6 @@ function scrollElementIntoView(id) {
 			block: 'center'
 		});
 	} catch (error) {
-		console.warn('scrollElementIntoView error:', error);
+		console.warn(error);
 	}
 }