+++ src/utils/stores.js
@@ -55,7 +55,8 @@ let __currentPage,
 	__homepageExtrasPanelVisible,
 	__wideWesbiteLayoutEnabled,
 	__signLanguageModeEnabled,
-	__verseWordBlocks;
+	__verseWordBlocks,
+	__offlineModeSettings;
 
 if (browser) {
 	const userSettings = JSON.parse(localStorage.getItem('userSettings'));
@@ -211,6 +212,9 @@ if (browser) {
 
 	// to store the visibility state of word blocks per verse
 	__verseWordBlocks = writable({});
+
+	// to store all the offline mode settings
+	__offlineModeSettings = writable(userSettings.offlineModeSettings);
 }
 
 export {
@@ -268,5 +272,6 @@ export {
 	__homepageExtrasPanelVisible,
 	__wideWesbiteLayoutEnabled,
 	__signLanguageModeEnabled,
-	__verseWordBlocks
+	__verseWordBlocks,
+	__offlineModeSettings
 };